Created Visuals for Initial Structure and Layout Window
This is the window layout without any of the functionality implemented yet. This component is currently designed to be displayed inside of a window or popup on another page, but it can be modified to fill the screen instead.

The reliability constraints button currently has a mockup drop down button functionality, which only serves as a test of what the real button will do.